Problem: Plugins are insufficiently namespaced
Solution: Prefix local declarations with the github username

Ideally we'd only load plugins live, but that brings the problems
of discovering and loading plugins. I'd like if, for example, the
ping hack necessary to make ping work on gokrazy only ran if people
used the vaelatern-ping plugin, but that's difficult to do unless we
dynamically load these. The go reflection package is interesting for
this, and may be able to dynamically cobble together a namespace... if
they adhere to a single function standard.

All in all, this approach of doing the plugins
works for now.  Eventually they might be namespaced as
github.com/Vaelatern/gokrazy-statuspage/internal/metrics/vaelatern-port-open
for example, if we can make that package layout actually work. This
would also allow greater complexity in the tests

I like disencouraging complexity in tests, since then we might need
to cache them, or perhaps the plugin itself would need to poll on
a schedule and cache the results itself, with the callback function
reading from the cache...